Add support for registration by leveraging a "Deferred" function
This reintroduces the possiblty to register a Gitlab Runner by Puppet
without the need of a manual step. This is done by using a "Deferred"
function which does the registration and returns the authentication
token for the configuration.

As Deferred is a >= Puppet 6 functionality this also drops support for
old Puppet versions.